move function to conftest
